requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
anyio==4.12.1
attrs==25.4.0
certifi==2026.1.4
discord.py==2.6.4
frozenlist==1.8.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
multidict==6.7.1
numpy==2.4.2
-e git+ssh://git@github.com/SymphonicTone/osrs-dump-scanner.git@e227c4a9d7e65e8aa76531fe807302090782175d#egg=osrs_dump_scanner
packaging==26.0
pandas==3.0.0
pluggy==1.6.0
propcache==0.4.1
Pygments==2.19.2
pytest==9.0.2
pytest-asyncio==1.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
respx==0.22.0
six==1.17.0
typing_extensions==4.15.0
yarl==1.22.0
